pyproject.toml
[project]
name = "planetas"
version = "0.1.0"
description = "Planetary zodiac position calculator"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
dependencies = [
    "pyswisseph>=2.10.3",
    "skyfield>=1.45",
    "click>=8.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
planetas = "planetas.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
